--- tolerance-policies/solutions/info/src/main/java/com/redhat/training/FaultsResource.java
@@ -1,27 +1,0 @@
-package com.redhat.training;
-
-import javax.inject.Inject;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-
-import com.redhat.training.services.InfoService;
-
-
-@Path("/faults")
-public class FaultsResource {
-
-    @Inject
-    InfoService infoService;
-
-    @POST
-    @Path("/crash")
-    public String activateCrash() {
-        return "'crash' has been set to true";
-    }
-
-    @POST
-    @Path("/reset")
-    public String deactivateCrashes() {
-        return "'crash' has been set to true";
-    }
-}
