+++ tolerance-policies/solutions/info/src/main/java/com/redhat/training/cpu/CpuStats.java
@@ -1,4 +1,4 @@
-package com.redhat.training.services;
+package com.redhat.training.cpu;
 
 import java.util.List;
