+++ tolerance-policies/solutions/info/src/main/java/com/redhat/training/sysinfo/InfoService.java
@@ -1,16 +1,18 @@
-package com.redhat.training.services;
+package com.redhat.training.sysinfo;
 
 import io.quarkus.logging.Log;
 import javax.enterprise.context.ApplicationScoped;
 import org.eclipse.microprofile.faulttolerance.Retry;
 
-
+/**
+ * Simulate a microservice that reads sytem info from a cloud instance
+ */
 @ApplicationScoped
 public class InfoService {
 
     private int callCount = 0;
 
-    @Retry(maxRetries = 5)
+    @Retry( maxRetries = 5 )
     public Info getInfo() {
         callCount++;
 
@@ -22,7 +24,7 @@ public class InfoService {
     private void crashPossibly() {
         if ( callCount % 5 > 0 ) {
             Log.error( "Request #" + callCount + " to InfoService has failed" );
-            throw new RuntimeException( "InfoService failed due to unexpected error" ); 
+            throw new RuntimeException( "InfoService failed due to unexpected error" );
         } else {
             Log.info( "Request #" + callCount + " to InfoService has succeeded" );
         }
